Read state of charge from CAN bus
This adds the much-awaited functionality to read messages on the CAN
bus. We're configuring our Orion to send state-of-charge as byte 4 of a
message with ID 0x6b0. It sends it as twice-SoC percentage (as an int).

I still need to confirm on the actual hardware that this properly
filters only packets with a specific ID. If it doesn't, I can mess with
the filters and/or add a software-level check.

+++ BatterySpoof/gauge.cpp
@@ -2,7 +2,7 @@
 
 analogWave wave(DAC);
 
-void set_gauge_soc(int soc) {
+void set_gauge_soc(float soc) {
   // ~47 Hz is the bottom of the SoC gauge, and 150 Hz is the top
   float freq = 47 + (1.03 * soc);
   wave.square(round(freq));
